Add API endpoint for asset file uploads with png and jpeg validation
Implements a /upload endpoint which uses the uploadAssetFile middleware to handle uploads.
This ensures any provided file is a png or jpeg and less than 1MB, and triggers and appropriate error response otherwise.

No actual parsing or handling of assets has been implemented yet - currently serves as a placeholder for future logic